[ir] Separate LaunchContextBuilder from Kernel
[ghstack-poisoned]
