+++ taichi/program/function.cpp
@@ -8,6 +8,7 @@ namespace taichi::lang {
 Function::Function(Program *program, const FunctionKey &func_key)
     : func_key(func_key) {
   this->program = program;
+  arch = program->compile_config().arch;
 }
 
 void Function::set_function_body(const std::function<void()> &func) {
